+++ css-classes.js
@@ -1,0 +1,90 @@
+import hashwasm from "hash-wasm"
+import fs from "fs"
+import postcss from "postcss"
+import stringify from "fast-stable-stringify"
+
+
+// This'll generate an Elm module with a function for each CSS class we have.
+// It will also generate a "CSS table" with the "css_class <=> elm_function" relation.
+// This "CSS table" makes it possible to only keep the CSS that's actually used.
+const plugin = async (flags, root, result) => {
+
+  const functions = []
+  const lookup = {}
+
+  let lastCls
+
+  root.walkRules(rule => {
+    if (!rule.selector.startsWith(".")) return;
+    if (rule.selector.includes(" ")) return;
+
+    const cls = rule
+      .selector
+      .split(",")[0]
+      .replace(/^\./, "")
+      .replace(/\\\./g, ".")
+      .replace(/\s?>\s?.*/, "")
+      .replace(/::.*$/, "")
+      .replace(/:not\([^\)]*\)/g, "")
+      .replace(
+        /(:(active|after|before|checked|disabled|focus|focus-within|hover|visited|nth-child\((even|odd)\)|(first|last)-child))+$/,
+        ""
+      )
+      .replace(/\\\//g, "/")
+      .replace(/\\([/])/g, "\\\\$1")
+      .replace(/\\([:])/g, "$1")
+
+    if (cls === lastCls) return;
+    lastCls = cls
+
+    const elmVariable = cls
+      .replace(/:/g, "__")
+      .replace(/__-/g, "__neg_")
+      .replace(/^-/g, "neg_")
+      .replace(/-/g, "_")
+      .replace(/\./g, "_")
+      .replace(/\//g, "over")
+
+    const elmVarWithProperCase = flags.elmNameStyle === "camel"
+      ? elmVariable.replace(/(_+\w)/g, g => g.replace(/_/g, "").toUpperCase())
+      : elmVariable
+
+    const css = rule
+      .toString()
+      .replace(/\s+/g, " ")
+      .replace(/(\w)\{/g, "$1 {")
+
+    functions.push(
+      `{-| This represents the \`.${cls}\` class.\n` +
+      `\n    ${css}` +
+      `\n-}\n` +
+      `${elmVariable} : Html.Attribute msg\n` +
+      `${elmVariable} = A.class "${cls}"\n`
+    )
+
+    lookup[elmVariable] = cls
+  })
+
+  const tmpDir = `./elm-stuff/elm-tailwind-css/${flags.elmModule}`
+  fs.mkdirSync(tmpDir, { recursive: true })
+
+  const header = `module ${flags.elmModule} exposing (..)\n\n`
+  const imports = [ "import Html", "import Html.Attributes as A" ]
+  const contents = header + imports.join("\n") + "\n\n" + functions.join("\n\n")
+  const table = stringify(lookup)
+  const hash = await hashwasm.xxhash32(table, 1)
+  const previousHash = fs.readFileSync(`${tmpDir}/css-table.cache`, { flag: "a+", encoding: "utf-8" })
+
+  if (hash === previousHash) return;
+
+  fs.writeFileSync(`${tmpDir}/css-table.cache`, hash)
+  fs.writeFileSync(`${tmpDir}/css-table.json`, table)
+  fs.writeFileSync(flags.elmPath, contents)
+
+}
+
+
+export default flags => ({
+  postcssPlugin: "elm-css-classes",
+  Once (root, { result }) { return plugin(flags, root, result) }
+})